backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
requests==2.32.3
psycopg[binary]==3.3.2
sqlalchemy==2.0.36
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
